backend/requirements.txt
flask
flask-cors
gunicorn
spacy
supabase
python-Levenshtein
PyPDF2
python-docx
werkzeug
  
